pyproject.toml
[project]
name = "next-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.5",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.1",
]
